First tests for chapter 5.
If I were writing this in Javascript, I would've used function currying to
separate the production die-roller from the version I use in tests. Common
Lisp seems to prefer some other approach (maybe with a macro?) but I don't
know enough yet to know the right way to solve this.